move any plotrelated functions to its own file
